Fix signed conversion warnings in Packet.cpp and ReadOnlyPacket.cpp that cause the build to fail on freebsd when -Wall / -Werror is enabled.

+++ libcomp/src/Packet.cpp
@@ -344,7 +344,7 @@ void Packet::WriteS16Big(int16_t value)
 void Packet::WriteS16Little(int16_t value)
 {
     // Convert the value to little endian byte order and then write it.
-    WriteS16(htole16(value));
+    WriteS16((int16_t)htole16(value));
 }
 
 void Packet::WriteU32(uint32_t value)
@@ -386,7 +386,7 @@ void Packet::WriteS32Big(int32_t value)
 void Packet::WriteS32Little(int32_t value)
 {
     // Convert the value to little endian byte order and then write it.
-    WriteS32(htole32(value));
+    WriteS32((int32_t)htole32(value));
 }
 
 void Packet::WriteU64(uint64_t value)
@@ -428,7 +428,7 @@ void Packet::WriteS64Big(int64_t value)
 void Packet::WriteS64Little(int64_t value)
 {
     // Convert the value to little endian byte order and then write it.
-    WriteS64(htole64(value));
+    WriteS64((int64_t)htole64(value));
 }
 
 void Packet::WriteFloat(float value)
